\name{mdep-package}

\alias{mdep-package}

\alias{mdep}

\docType{package}

\title{
midas data extraction and pre-processing
}

\description{
Few (very basic!) functions on the data extraction and pre-processingof traffic counting data 
of the Highways England MIDAS network. These function are used to prepare the road networks
and traffic data sets for macroscopic traffic flow models (such as CTM/METANET).
To use these functions you need to have access to the MIDAS, which is available upon request.
Raw data/transformed data are explained/provided in my PhD thesis.
}

\details{
	\tabular{ll}{
	Package: \tab mdep\cr
	Type: \tab Package\cr
	Version: \tab 1.0\cr\cr
	Date: \tab 2016-11-05\cr
	License: \tab GPL (>=2) \cr
	}
}

\author{
Arwa Sayegh

Maintainer: Arwa Sayegh <ts11ass@leeds.ac.uk>
}